New structure for color type
